Code slightly dryer by taking getCode() into footer rather than gamePage
