+++ src/client/app/containers/FooterContainer.jsx
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 // import createCommand from '../actions/UserCommandAction.js';
 import $ from 'jquery';
 import changeLevel from '../actions/changeLevel.js';
+import setCode from '../actions/setCode.js';
 import createGame from '../game/game.js';
 
 class Footer extends Component {
@@ -34,8 +35,51 @@ class Footer extends Component {
     }
 
     /* Fetches and renders the user's code for that level into the editor */
-    this.props.getCode();
+    // this.props.getCode();
 
+
+
+    this.getCode();
+
+  }
+
+  getCode() {
+    /* Fetches the user's solution code from the database and     */
+    /* sends the code to the console so the code editor will      */
+    /* render with the appropriate text                           */
+    var url = `/game?username=${ this.props.user }`;
+
+    fetch(url, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    }).then(response => {
+      response.json().then(solutions => {
+
+        /** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ***/
+
+        /** THIS MAY HAVE BEEN FIXED, I'M NOT SURE                                                  **/
+        /** Throws an error when it tries to fetch code for a user that has never saved code before **/
+        /** This problem may or may not be present anymore... Troubleshooting required              **/ 
+        /** The reason is most likely that getCode() was being run when it shouldn't be? Test it    **/
+
+        /** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ***/
+
+        var result = solutions.filter(solution => {
+          return solution.level === this.props.level;
+        })[0];
+        var solution = result ? result.solution : '// Input your code here\n\n';
+
+        /* Update the current code in the Redux Store */
+        this.props.setCode(solution);
+
+        // return solution for promise chaining
+        return solution;
+      })
+    }).catch(err => {
+      console.log('Error fetching solution code: ', err);
+    })
   }
 
   render() {
@@ -65,6 +109,7 @@ class Footer extends Component {
 
 var mapStateToProps = state => {
   return {
+    user: state.user,
     level: state.level
   }
 }
@@ -73,6 +118,11 @@ var mapDispatchToProps = dispatch => {
   return {
     changeLevel: (level) => {
       dispatch(changeLevel(level));
+    },
+    setCode: (code) => {
+      /* Function to dispatch an action that will set the current code in the Redux Store             */
+      /* This is used for making sure the code in the Redux Store matches with the code in the editor */
+      dispatch(setCode(code));
     }
   }
 }
