+++ src/client/views/GamePage.jsx
@@ -4,7 +4,6 @@ import { withRouter } from 'react-router';
 import Game from '../app/components/Game.jsx';
 import Console from '../app/containers/ConsoleContainer.jsx';
 import Footer from '../app/containers/FooterContainer.jsx';
-import setCode from '../app/actions/setCode.js';
 
 /*******************************************************/
 /* Renders the game page that consists of:             */
@@ -25,50 +24,7 @@ class GamePage extends Component {
     /* tries to access this page by typing in the url             */
     if (!this.props.user) {
       this.props.router.push('/');
-    } else {
-      /* If there is a valid user logged in, then fetch the       */
-      /* user's solution code from the database                   */
-      this.getCode();
-    }
-  }
-
-  getCode() {
-    /* Fetches the user's solution code from the database and     */
-    /* sends the code to the console so the code editor will      */
-    /* render with the appropriate text                           */
-    var url = `/game?username=${ this.props.user }`;
-
-    fetch(url, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json'
-      }
-    }).then(response => {
-      response.json().then(solutions => {
-
-        /** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ***/
-
-        /** THIS MAY HAVE BEEN FIXED, I'M NOT SURE                                                  **/
-        /** Throws an error when it tries to fetch code for a user that has never saved code before **/
-        /** This problem may or may not be present anymore... Troubleshooting required              **/ 
-        /** The reason is most likely that getCode() was being run when it shouldn't be? Test it    **/
-
-        /** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ** ERROR ***/
-
-        var result = solutions.filter(solution => {
-          return solution.level === this.props.level;
-        })[0];
-        var solution = result ? result.solution : '// Input your code here\n\n';
-
-        /* Update the current code in the Redux Store */
-        this.props.setCode(solution);
-
-        // return solution for promise chaining
-        return solution;
-      })
-    }).catch(err => {
-      console.log('Error fetching solution code: ', err);
-    })
+    } 
   }
 
   render() {
@@ -78,7 +34,7 @@ class GamePage extends Component {
       <div>
         <Game level={ this.props.level } />
         <Console />
-        <Footer getCode={ this.getCode.bind(this) } />
+        <Footer/>
       </div>
     ) : null
   }
@@ -94,12 +50,8 @@ const mapStateToProps = state => {
 }
 
 const mapDispatchToProps = dispatch => {
-  /* Function to dispatch an action that will set the current code in the Redux Store             */
-  /* This is used for making sure the code in the Redux Store matches with the code in the editor */
   return {
-    setCode: (code) => {
-      dispatch(setCode(code));
-    }
+
   }
 }
